+++ changelog.md
@@ -1,2 +1,3 @@
-+ [BUG] : Fix name of i18n translate function
-+ [BUG] : Fix name of {create,remove}_dirs in utils
++ [BUG] : Fix name of i18n translate function.
++ [BUG] : Fix name of {create,remove}_dirs in utils.
++ [BUG] : Fix some references.